reorganise dashboard routes
